Fix a warning.
Commit migrated from https://github.com/mono/mono/commit/a5101bb67ee6a3f106f9a0e6e0e9e7000b47293c

+++ src/mono/mono/metadata/socket-io.c
@@ -2326,7 +2326,7 @@ ves_icall_System_Net_Sockets_Socket_WSAIoctl (SOCKET sock, gint32 code,
 
 	*error = 0;
 	
-	if (code == FIONBIO) {
+	if ((guint32)code == FIONBIO) {
 		/* Invalid command. Must use Socket.Blocking */
 		return -1;
 	}
